adição de arquivos

+++ loja/include/bebida.h
@@ -1,22 +1,32 @@
+/**
+* @file bebida.h
+* @brief Cabeçalho da subclasse bebida
+* @author Antonio Marcos Oliveira
+* @since 13/05/2018
+* @date 17/05/2018
+*/
 #ifndef _BEBIDA_H_
 #define _BEBIDA_H_
 
 #include "produto.h"
 
+/**
+ * @brief      Class for bebida.
+ */
 class Bebida : public Produto
 {
 public:
 	Bebida();
-	Bebida(string _codigo, string _descricao, short _preco, 
-		  float _teor_alcoolico);
+	Bebida(string _codigo, string _descricao, float _preco, 
+		  int _teor_alcoolico);
 	~Bebida();
 private:
-	float m_teor_alcoolico;
+	int m_teor_alcoolico;
 public:
 	// getters
-	float get_teor_alcoolico();
+	int get_teor_alcoolico();
 	// setters
-	void set_teor_alcoolico(float teor_alcoolico_);
+	void set_teor_alcoolico(int teor_alcoolico_);
 private:
 	ostream& print(ostream &o) const;
 	istream& print(istream &i) const;
